backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
astrapy
bcrypt
email-validator
requests
gunicorn
langchain
numpy
